chore: 1. formatting for remove the warning during compilation 2. simple_malloc to alloc.c 3. refactoring the -D flags
