+++ lab03/kernel/include/alloc.h
@@ -1,0 +1,9 @@
+#ifndef __ALLOC_H__
+#define __ALLOC_H__
+
+#include "type.h"
+
+void  mem_init();
+void* simple_malloc(uint32_t size);
+
+#endif
