+++ lab03/kernel/include/lib.h
@@ -6,7 +6,4 @@
 uint32_t strtol(const char *sptr, uint32_t base, int size);
 int atoi(const char *str);
 
-void  mem_init();
-void* simple_malloc(uint32_t size);
-
 #endif
