+++ lab03/kernel/src/shell.c
@@ -4,7 +4,7 @@
 #include "reboot.h"
 #include "io.h"
 #include "string.h"
-// #include "cpio.h"
+#include "alloc.h"
 #include "lib.h"
 #include "timer.h"
 #include "irq.h"
@@ -14,7 +14,6 @@ static void help(int argc, char *argv[]);
 static void mailbox(int argc, char *argv[]);
 static void test_malloc(int argc, char *argv[]);
 static void reboot(int argc, char *argv[]);
-static void print_time(int argc, char *argv[]);
 static void set_timeout(int argc, char *argv[]);
 
 extern void cpio_list(int argc, char *argv[]);
@@ -33,7 +32,6 @@ cmd cmds[] =
     {.name = "malloc",  .func = &test_malloc,.help_msg = "\nmalloc\t: test malloc function"},
     {.name = "reboot",  .func = &reboot,     .help_msg = "\nreboot\t: reboot the device"},
     {.name = "exec",    .func = &cpio_exec,  .help_msg = "\nexec\t: execute a file in the cpio archive"},
-    // {.name = "print_time", .func = &print_time, .help_msg = "\nprint_time\t: print the time after booting and set 2 sec timeout"},
     {.name = "async",   .func = &async_shell, .help_msg = "\nasync\t: enter dummy async shell mode"},
     {.name = "set_timeout", .func = &set_timeout, .help_msg = "\nset_timeout\t: set a new timer"}
 };
@@ -75,7 +73,6 @@ void async_shell()
         async_uart_puts(command);        
         if(strcmp(command, "exit") == 0)
         {
-            // async_uart_puts("\r\nExiting async shell...");
             printf("\r\nExiting async shell...");
             break;
         }
@@ -130,11 +127,6 @@ static void test_malloc(int argc, char **argv)
     }
 }
 
-static void print_time(int argc, char **argv)
-{
-    core_timer_enable();
-    // set_timer(2);
-}
 
 static void set_timeout(int argc, char *argv[])
 {
@@ -142,9 +134,7 @@ static void set_timeout(int argc, char *argv[])
         printf("\nUsage: set_timeout <msg> <sec>");
         return;
     }
-    // printf("\nSetting timeout...");
     char* msg = argv[1];
-    // printf(argv[2]);
     unsigned int sec = atoi(argv[2]);
     add_timer(print_timeout_msg, msg, sec);
     core_timer_enable();
@@ -186,7 +176,6 @@ int split_command(char* command, char *argv[])
     while(*token != '\0')
     {
         argv[argc] = token;
-        // printf(token);
         argc++;
         token = strtok(0, " ");
     }
